añadi la opcion de votar por un candidato
